[Brief] generate Brief when user wants to share by email
